added animalData Type to useState in all levels
