+++ frontend/src/context/LevelProvider.tsx
@@ -3,11 +3,15 @@ import {createContext, ReactElement, useState} from "react";
 export interface LevelContextType {
     levelOfPlayer?: number
     setNewlevelOfPlayer: (newLevel: number) => void
+    levelUp: () => void
 }
 
 export const LevelContext = createContext<LevelContextType>({
     setNewlevelOfPlayer: () => {
         throw Error("defaut setNewLevel function has not been initialized")
+    },
+    levelUp: () => {
+        throw Error("defaut levelUp function has not been initialized")
     }
 })
 
@@ -15,12 +19,21 @@ export default function LevelProvider({children}: { children: ReactElement<any, 
 
     const [levelOfPlayer, setLevelOfPlayer] = useState<number>(1)
 
+    const levelUp = () => {
+        if (levelOfPlayer === undefined) {
+            setNewlevelOfPlayer(1)
+        } else {
+            const newLevel: number = levelOfPlayer + 1;
+            setNewlevelOfPlayer(newLevel)
+        }
+    }
+
     const setNewlevelOfPlayer = (newLevel: number) => {
         setLevelOfPlayer(newLevel)
     }
 
     return (
-        <LevelContext.Provider value={{levelOfPlayer: levelOfPlayer, setNewlevelOfPlayer: setNewlevelOfPlayer}}>
+        <LevelContext.Provider value={{levelOfPlayer: levelOfPlayer, setNewlevelOfPlayer: setNewlevelOfPlayer, levelUp:levelUp}}>
             {children}
         </LevelContext.Provider>
     )
